server/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Pillow==10.4.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==23.0.0
openai==1.72.0
nltk==3.8.1